requirements.txt
#requirements.txt

Flask==2.0.1
gunicorn==20.1.0
Werkzeug==2.0.3
google-cloud-bigquery==2.34.0
google-cloud-storage==2.17.0